container: move Container to container.go
