+++ container/container.go
@@ -2,6 +2,7 @@ package container
 
 import (
 	"context"
+	"io"
 	"time"
 
 	"github.com/docker/docker/api/types"
@@ -10,6 +11,25 @@ import (
 	docker "github.com/docker/docker/client"
 	"github.com/mesg-foundation/core/config"
 )
+
+// Container describes the API of container package.
+type Container interface {
+	Build(path string) (tag string, err error)
+	CreateNetwork(namespace []string) (id string, err error)
+	DeleteNetwork(namespace []string, event EventType) error
+	FindContainer(namespace []string) (types.ContainerJSON, error)
+	FindNetwork(namespace []string) (types.NetworkResource, error)
+	FindService(namespace []string) (swarm.Service, error)
+	ListServices(labels ...string) ([]swarm.Service, error)
+	ListTasks(namespace []string) ([]swarm.Task, error)
+	Namespace(ss []string) string
+	ServiceLogs(namespace []string) (io.ReadCloser, error)
+	SharedNetworkID() (networkID string, err error)
+	StartService(options ServiceOptions) (serviceID string, err error)
+	Status(namespace []string) (StatusType, error)
+	StopService(namespace []string) (err error)
+	TasksError(namespace []string) ([]string, error)
+}
 
 // DockerContainer provides high level interactions with Docker API for MESG.
 type DockerContainer struct {
